chore: Move the psutil mock into a test-only package
The mock is a regular build file, so testify and its dependencies are linked
into the shipped binary and their init functions run on every start. It cannot
just be renamed to a test file, as MockPS, MockPSDisk and MockDiskUsage are
used by the tests of seven other plugins.

Moving it to the psutiltest package, which nothing outside of tests imports,
keeps the mock available across package boundaries while dropping the testing
dependency of the psutil package.
