+++ plugins/inputs/diskio/diskio_test.go
@@ -7,7 +7,7 @@ import (
 	"github.com/shirou/gopsutil/v4/disk"
 	"github.com/stretchr/testify/require"
 
-	"github.com/influxdata/telegraf/plugins/common/psutil"
+	"github.com/influxdata/telegraf/plugins/common/psutil/psutiltest"
 	"github.com/influxdata/telegraf/testutil"
 )
 
@@ -129,7 +129,7 @@ func TestDiskIO(t *testing.T) {
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			var mps psutil.MockPS
+			var mps psutiltest.MockPS
 			mps.On("DiskIO").Return(tt.result.stats, tt.result.err)
 
 			var acc testutil.Accumulator
@@ -189,7 +189,7 @@ func TestDiskIOUtil(t *testing.T) {
 	}
 
 	var acc testutil.Accumulator
-	var mps psutil.MockPS
+	var mps psutiltest.MockPS
 	mps.On("DiskIO").Return(cts, nil)
 	diskio := &DiskIO{
 		Log:     testutil.Logger{},
@@ -202,7 +202,7 @@ func TestDiskIOUtil(t *testing.T) {
 	// sleep
 	time.Sleep(1 * time.Second)
 	// gather twice
-	mps2 := psutil.MockPS{}
+	mps2 := psutiltest.MockPS{}
 	mps2.On("DiskIO").Return(cts2, nil)
 	diskio.ps = &mps2
 
@@ -242,7 +242,7 @@ func TestCounterWraparound(t *testing.T) {
 	}
 
 	var acc testutil.Accumulator
-	var mps psutil.MockPS
+	var mps psutiltest.MockPS
 	mps.On("DiskIO").Return(cts, nil)
 	diskio := &DiskIO{
 		Log:     testutil.Logger{},
@@ -255,7 +255,7 @@ func TestCounterWraparound(t *testing.T) {
 	require.NoError(t, diskio.Gather(&acc))
 
 	// Second gather with wrapped counters
-	mps2 := psutil.MockPS{}
+	mps2 := psutiltest.MockPS{}
 	mps2.On("DiskIO").Return(cts2, nil)
 	diskio.ps = &mps2
